requirements.txt
azure-mgmt-resource >= 20.1.0
azure-mgmt-compute >= 25.0.0
azure-mgmt-network >= 19.3.0
azure-identity >= 1.7.1